[MAT-196] Make arguments of expressions immutable.
This is done using the PtrVector class, which holds an STL vector
but only permits its contents to be read, not modified, and does
not allow null pointers to be inserted.

Presently only wrapper methods that are used are implemented.
Others may need to be implemented to get to other functionality of
the underlying vector in the future.

The consquence of making its contents immutable (and returning
only const pointers to its elements) is that dynamic_cast is not
possible because it would cast away the const-ness. In order to
get around this, a dynamic_cast-like mechanism is implemented for
the OGNumeric class.

So far this is only implemented for the scalar types: OGNumeric
has members asOG*Scalar, which are virtual and return null. The
derived class then overrides the method that mentions its name
(e.g. OGRealScalar overrides asOGRealScalar) so that it returns
the this pointer.

Thus, the effect of calling OGNumeric::asOGRealScalar() is the
same as using dynamic_cast<OGRealScalar*>(ptr), except that it
uses const pointers, and will work across module boundaries,
unlike C++ RTTI.