[MAT-199] Check for null args on construction.
Unfortunately there seems to be no way to get rid of OGExpr(),
because of the one- and two-argument constructors of OGUnaryExpr
and OGBinaryExpr. However, OGExpr objects will never end up with
null args as long as OGExpr subclass constructors always ensure
that setArgs() is called if the use the default OGExpr()
constructor.