[MAT-199] Remove noarg ctors, no nulls in OGExpr()
OGExpr() now sets _args to point to a new ArgContainer. setArgs()
results in the deletion of the old arguments so that they are not
leaked. It should now not be possible to create any expression
nodes with null arguments, and therefore it shouldn't be possible
to have expressions with invalid state.

Default constructors for other expression nodes are removed.