[MAT-240] Define containers in terms of held type
Declaring PtrVector<T> to get a container of const T* was
seriously confusing. This commit redefines it so that you instead
define PtrVector<const T*> to get a container of const T*.

This does not mean it will work for non ptr-to-const types. A
compile-time check for this is forthcoming.