[MAT-242] Generate report from GTest tests.
This was not trivial as expected. GTest generated JUnit output but
the output comes in one xml file for each test.

The XSL for generating an HTML report from JUnit XML files only
works on one file at a time. In order to generate a single report
from multiple files, Ant reads in an aggregates XML files on the
fly whilst generating the report.

The aggregate-xml.py script performs a similar task to generate a
single XML file that can be handed to xsltproc.