[MAT-248] Make sure every bind is unbound.
Critical sections were entered and never left because arrays were
being bound and not unbound. This is fixed by:

1. Deleting the chain after calling entrypt.
2. Fixing the resulting crash in the unbind by storing a jobject
in _backingObject instead of jobject*. The jobject* was previously
obtained as the address of a stack variable in createExpression,
which is an error.

A jobject is a typedef _jobject* so it is suitable for being held
in the _backingObject field.