[MAT-247] Add target to run Jacoco.
This invokes the tests with the Jacoco agent and then generates
a test report. It can be run with:

  make jacoco

The $JACOCOJAR environment variable must be set to tell CMake
where to find the Jacoco agent.

The report generation requires reporter-0.0.1-SNAPSHOT.jar,
org.jacoco.report-0.6.3.201306030806.jar,
org.jacoco.core-0.6.3.201306030806.jar, and asm-all-4.1.jar.