[MAT-255] Add materialisation to OGTerminal
This adds materialisation to OGTerminal types, for scalars, dense
and diagonal matrices.

toReal16Array and toComplex16Array methods are added to
OGTerminal, which are presently implemented for diagonal matrices
and are used to materialise them efficiently.

Functions for creating Java arrays of arrays are moved into
dispatch.cc, since they are reused by the dispatcher that
materialises OGTerminals. Methods for constructing Java arrays of
doubles are also added.

Librdag code coverage: 100%