[MAT-228] Double-negate and complex matrix tests.
This exposed a bug in the complex matrix negate runner, which was
returning the passed-in data instead of the computed data.

The matrix results are manually cleaned up because the negate
runners presently return terminals instead of registers. There is
no convenient way to turn a returned value into a register in
order to decrement the reference, so Registers are not used for
now.

Test functionality is factored into the run_entrypt function. This
saves code repetition but allows tests of different classes of
node to have different names.