[MAT-300] Add exceptions with stack trace info
This required exceptions to be split into a separate namespace,
dogma_exceptions, since it contains code common to librdag and
convert.

The exceptions use GNU execinfo to construct a backtrace and
record this in BacktraceElement objects, one for each frame.

The existing rdag_error and convert_error classes are now direct
subclasses of dogma_error.

Some testing is added to check the exception classes where
possible. Since the stack can vary from one environment to the
next, only the topmost frames are examined for their function
names (addresses and filenames can vary from run to run). In the
case of assignment and copy, comparisons are made between the
original and new object to avoid this inconsistency between
machines/runs being a problem.