[MAT-300] Propagate exceptions/backtraces to Java
The MathsExceptionNative class inserts the backtrace passed to its
constructor by the JNI code into the Java stack trace. This is the
base for exceptions in native code. Conversion or computation
exceptions actuall instantiate MathsExceptionNativeConversion or
MathsExceptionNativeComputation.

The C++ exceptions are caught by try-catch blocks in the highest
level JNI methods, and convert any caught exception to the
appropriate corresponding Java exception.

This does not address release of resources in the presence of
exceptions - i.e leaks are still possible. Exception safety against
memory leaks remains to be addressed in future.