[MAT-337] Catch standard exceptions
This adds catching of standard exceptions to JShim. Since we only
call standard library functions from with JShim and RDAG, we now
catch all the exceptions that can be thrown.

Because standard exceptions do not carry backtrace information, we
can only translate the exception message into a Java exception and
throw it - it is not possible to provide additional information.

The Java-side MathsExceptionNativeUnspecified is used for the
standard exceptions, since we do not know which exception will
have been thrown, and it was not one that we specified.