[MAT-246] ArgContainers constructed by expressions
It is no longer necessary to construct an ArgContainer outside
an expression constructor. Instead, expressions take their actual
arguments and build the ArgContainer themselves, in order to
simplify memory management.

The use of a vector<const OGNumeric*>* for passing the results of
generateArgs() is a temporary measure, and will be removed when
the translation is converted to a procedural implementation.

Some missing exception checks are also added as they were spotted
during modification of the JNI interface.