[MAT-372] PtrVector: ownership a runtime property
OwningPtrVector and NonOwningPtrVector are collapsed into the
single PtrVector type. A PtrVector always begins life as a non-
owning entity. It can then be made to own its elements later on.

This is because it is almost never desirable to own elements
immediately, because some work may have to be done to populate it,
which could throw an exception.

Change-Id: I4a5f82c6af193dcd54d8ba41230b68eafffb14f7