[MAT-378] Version info in JAR, fix rebuilding
Rebuilding of the JAR after some a non-java file that is included
in it changes should now be fixed. This is achieved by always
rebuilding a JAR containing native code, by using a custom target
to build the JAR instead of a custom command.

Also, handling of the DEPENDS argument to add_jar() is implemented
 - a DEPENDS argument was specified in the main JAR's add_jar, but
DEPENDS was not actually being handled by add_jar.

A small script to generate a text file containing the SHA1 of the
commit for inclusion into the JAR is added. The correct updating
of this file every time the JAR is built is handled by using the
new DEPENDS argument.