[MAT-378] Add blob creation tool.
A blob consists of a set of artifacts and a verinfo file
describing them, wrapped up in a zip file. Blobs are used so that
artifacts can be uploaded with information about their build in a
single file, either to the buildmaster, or another machine being
used to host built code.