[MAT-378] Add tool to combine blobs
This tool takes a set of platform-specific blobs and creates a
platform-independent blob that consists of a platform independent
main jar and the tests, sources and javadoc jars.

Verinfo from each blob is combined to produce a single verinfo
file with information about the builds that the combined main jar
came from.