[MAT-387] Make terminal constructors private
Since it is a requirement that every OGNumeric has a shared_ptr
that it can use for shared_from_this(), we don't want to allow the
construction of an OGNumeric with a "naked" new.

This commit makes all OGTerminal and subclasses constructors
private, and provides a static factory method create() for their
construction. The factory method returns a new OGTerminal wrapped
in a shared_ptr.

This still leaves OGExpr constrcutors exposed; this is addressed
in the next commit.