[MAT-412] Keep _convertor private to  OGTerminal
This prevents outside objects potentially getting access to a
dangling ConvertTo pointer (which they do not need anyway). In
order to implement this chage, convertto.hh needed to not depend
on terminal.hh, to break a cyclic dependency. In order to achieve
this, some forward declarations are added to convertto.hh