[MAT-180] Attempt to replace all sizes with size_t
This involves those values not directly touched by the JVM.

Because all sizes of things are replaced with size_t, it is not
possible for negative sizes to be used any more - the code that
checks for this and the tests of this check are removed.

However, the fields holding sizes are signed in Java. To ensure
that the passing of sizes from Java to C++ is safe, we replace the
getIntFromVoidJMethod with getSizeTFromVoidJMethod that checks
that the received value is non-negative, before casting it to
size_t.

This should leave the remaining ints (at least in librdag) all
safe to convert to int4.