[MAT-420] Switched C++ to use formal name OGDenseComplexMatrix
Very similar to previous commit, but now in complex space.

Java and C++ now have the same names for the same types.
This was done largely with:
find . -type f -print0 |\
xargs -0 sed -i 's/OGComplexMatrix/OGComplexDenseMatrix/g'

and similar for the EXPR_ENUM enum equivalent. A minor patch up
was needed in the python code that generates the dispatch mech.