[MAT-437] translated java mldivide to C++
This commit contains a very crude and somewhat literal translation
of the legacy,
MldivideOGMatrixOGMatrix.java
and
MldivideOGMatrixOGMatrixHelper.java
classes.

Commit is purely work in progress. The code currently leaks and
unit tests have input data but no expected result.

In writing this it has become apparent that we need an
"unrecoverable_error" type exception which does what it suggests
and can be extended to e.g. "xerbla_error". This way we can just
catch recoverable errors and let the rest just propagate.

This is going to be a long haul...