[MAT-441] fix up loop again, need 0 included.
This fixes a rather fiddly walk through the pinv() singular
values. The loop needs to include hitting zero *and* needs to
carry a flag to "skip" if non values within threshold are found.
Using an int64_t for this, negative value indicates skip and we
get zero access correctly too.