[MAT-443] fix segfault in mldivide from "triangle" pass.
The fuzzer picked up a segfault in mldivide for the data presented
in the regression test included. Essentially, a row containing all
zeros, which is a valid input, could cause a segfault due to the
value -1 being a flag and it not being overwritten by a rowStart
as there effectively isn't one in this case. So a lookup of index
size_t = -1 occurs and causes havoc. Patch basically checks indices
are valid before attempting the check of permutation validity.