[MAT-444] remove dynamic libstdc++
The dynamic linking of libstdc++ is swapped for static linking.
This is done, in the parlance of GCC, to "freeze" the version of
libstc++ against which the build links. This is so that C++
symbols are guaranteed to resolve correctly regardless of other
libstdc++ versions potentially loaded in the ld cache.