[MAT-450] Add manifests to tests on Windows
Installer Detection Technology on 32-bit Windows mistakes many
tests for installers and pops up a UAC prompt when they are
executed.

This commit adds a manifest to test programs to prevent this
happening. This applies to both 32- and 64-bit Windows, as it is
likely that future versions of Windows will require a manifest to
be present.