[MAT-465] Differentate between toString() and toDebugString().
This switches all the current toString() methods to toDebugString(),
it then adds cleaner toString() methods.
The implementation is forced via the addition of an abstract method
to the OGTerminal base class. The reason for this switch is that
users are unlikely to benefit from seeing the debug information in
standard operation. Both are therefore implmented and the default is
to now print something more useful to users.