[MAT-462] Add new exception types and make use of them.
librdag gets two new exception types:

librdag_unrecoverable_exception (thrown in the case of an
unrecoverable but non fatal execution path).

librdag_recoverable_exception (thrown in the case of something
that might be objectionable happening, but it is recoverable).

These two exceptions are applied to LAPACK xgesvd as proof of
concept.