[MAT-462] respond to PR comments. Fix up qNaN and try...catch.
This patch is in response to PR #128 comments. It:

Adds in demo/tests for librdag::isfinite() correctly reporting
that a quiet NaN is indeed not finite.

Reworks the exception throwing chains in a number of runners so
that the only exception caught is of the rdag_recoverable_error
type. All other exceptions should rightly propagate from these
call sites.