[MAT-479] Add requirement for finite system to mldivide.
This adds a scan of the system matrix to the mldivide function.
This scan will cause an exception to be raised if the matrix
contains non-finite data. The reason for this behavior is to
prevent LAPACK solvers receiving data containing inf/NaN as it
makes them do rightly bizarre things.